app_metrics/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-auth
